Refactor column definitions to reduce repeated code
