+++ src/modules/audit/components/ClientAuditHistory.tsx
@@ -1,20 +1,13 @@
-import { Paper, Stack, Typography } from '@mui/material';
-import { compact, filter } from 'lodash-es';
-import AuditObjectChangesSummary, {
-  ObjectChangesType,
-} from './AuditObjectChangesSummary';
-import {
-  ContextualCollapsibleList,
-  ContextualCollapsibleListsProvider,
-  ContextualListExpansionButton,
-} from '@/components/elements/CollapsibleListContext';
+import { Paper } from '@mui/material';
+import { ContextualCollapsibleListsProvider } from '@/components/elements/CollapsibleListContext';
 import { ColumnDef } from '@/components/elements/table/types';
 import PageTitle from '@/components/layout/PageTitle';
 import useSafeParams from '@/hooks/useSafeParams';
+import {
+  AuditHistoryNode,
+  auditHistoryColumns,
+} from '@/modules/audit/components/auditHistoryColumnDefs';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
-import { hasMeaningfulValue } from '@/modules/form/util/formUtil';
-import RelativeDateTableCellContents from '@/modules/hmis/components/RelativeDateTableCellContents';
-import { auditActionForDisplay } from '@/modules/hmis/hmisUtil';
 import {
   ClientAuditEventFilterOptions,
   GetClientAuditEventsDocument,
@@ -22,76 +15,7 @@ import {
   GetClientAuditEventsQueryVariables,
 } from '@/types/gqlTypes';
 
-type AuditHistoryType = NonNullable<
-  NonNullable<GetClientAuditEventsQuery['client']>['auditHistory']
->['nodes'][0];
-
-const columns: ColumnDef<AuditHistoryType>[] = [
-  {
-    header: 'Timestamp',
-    width: '180px',
-    render: (e) => (
-      <RelativeDateTableCellContents dateTimeString={e.createdAt} />
-    ),
-  },
-  {
-    header: 'User',
-    width: '180px',
-    render: ({ user, trueUser }) =>
-      compact([trueUser?.name, user?.name]).join(' acting as ') ||
-      'System User',
-  },
-  {
-    header: 'Action',
-    width: '100px',
-    render: ({ event }) => auditActionForDisplay(event),
-  },
-  {
-    header: 'Record Type',
-    width: '150px',
-    render: ({ recordName, recordId }) => {
-      return (
-        <Stack>
-          <Typography variant='inherit'>{recordName}</Typography>
-          <Typography
-            color='text.secondary'
-            variant='inherit'
-          >{`ID: ${recordId}`}</Typography>
-        </Stack>
-      );
-    },
-  },
-  {
-    header: (
-      <Stack direction='row' justifyContent='space-between' alignItems='center'>
-        <strong>Fields Changed</strong>
-        <ContextualListExpansionButton />
-      </Stack>
-    ),
-    tableCellProps: {
-      sx: { p: 0, backgroundColor: (theme) => theme.palette.grey[50] },
-    },
-    render: (e) => {
-      if (!e.objectChanges || Object.keys(e.objectChanges).length === 0) {
-        return null;
-      }
-
-      const labels = Object.values(e.objectChanges as ObjectChangesType)
-        .filter((r) => filter(r.values, hasMeaningfulValue).length > 0)
-        .map((val) => val.displayName);
-
-      return (
-        <ContextualCollapsibleList title={labels.join(', ')}>
-          <AuditObjectChangesSummary
-            objectChanges={e.objectChanges as ObjectChangesType}
-            recordType={e.graphqlType}
-            eventType={e.event}
-          />
-        </ContextualCollapsibleList>
-      );
-    },
-  },
-];
+const columns: ColumnDef<AuditHistoryNode>[] = auditHistoryColumns;
 
 const ClientAuditHistory = () => {
   const { clientId } = useSafeParams() as { clientId: string };
@@ -103,7 +27,7 @@ const ClientAuditHistory = () => {
         <GenericTableWithData<
           GetClientAuditEventsQuery,
           GetClientAuditEventsQueryVariables,
-          AuditHistoryType,
+          AuditHistoryNode,
           ClientAuditEventFilterOptions
         >
           columns={columns}
